Added column filter feature
Added a feature that allows the columns that get written to the csv to be controlled by a section in the ini file. This get read into a dictionary which is looped through and used to drop unwanted columns for the pandas dataframe before the data is written to the CSV file.